pyproject.toml
[project]
name = "DSA5104Proj_StackExchange"
version = "0.1.0"
description = ""
authors = [
    {name = "ZZDirty", email = "muztaga2@gmail.com"},
]
dependencies = [
    "pymongo[srv]>=4.5.0",
    "tqdm>=4.66.1",
    "ipykernel>=6.26.0",
    "numpy>=1.26.1",
    "pandas>=2.1.2",
    "matplotlib>=3.8.1",
    "scikit-learn>=1.3.2",
    "seaborn>=0.13.0",
    "mysql-connector-python>=8.2.0",
]
requires-python = ">=3.9,<3.11"
readme = "README.md"
license = {text = "MIT"}
